extract_transaction_metrics
